% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewind_version.R
\name{rewind_version}
\alias{rewind_version}
\title{Make changes after a version has been finalized}
\usage{
rewind_version(major, minor, path = "version.rds")
}
\arguments{
\item{major}{(numeric) the current major version of your project}

\item{minor}{(numeric) the current minor version of your project}

\item{path}{a string indicating where \code{version.rds} is stored. This
should \emph{always} be stored in the project's main directory.}
}
\value{
Nothing is returned. The version in \code{version.rds} is updated to an
older version.
}
\description{
The main purpose of this function is to return to an earlier
version of a project in order to make changes. That is, it undoes the
result of \href{finalize_version}{}.
}
