% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_helpers.R
\name{use_helpers_tidyverse}
\alias{use_helpers_tidyverse}
\title{Add tidyverse helper functions to R/}
\usage{
use_helpers_tidyverse()
}
\value{
Nothing. Modifies your workspace.
}
\description{
Copies \code{summarize_each_group.R} into the \verb{R/} directory of the current
project. Contains \code{summarize_each_group()}, an extension of
\code{dplyr::summarize()} that computes a summary for each group as well as
an overall summary, stacking the results into a single data frame. Prints
\code{add_package()} calls for the packages these helpers depend on.
}
