% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pericircumference_wrappers.R
\name{use_pericircumference_report_qt}
\alias{use_pericircumference_report_qt}
\title{Initialize a Quarto HTML report project}
\usage{
use_pericircumference_report_qt()
}
\value{
Nothing. Modifies your workspace.
}
\description{
A wrapper around \code{\link[=use_pericircumference]{use_pericircumference()}} for experienced users who want
to quickly set up a project with a Quarto report and no slides or tutorials.
Equivalent to calling
\code{use_pericircumference(doc_format = "quarto", include_tutorials = FALSE, include_report = TRUE, include_slides = FALSE)}.
}
