% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package.R
\name{add_package}
\alias{add_package}
\title{Add a package to packages.R}
\usage{
add_package(name, purpose = NULL)
}
\arguments{
\item{name}{the name of the package}

\item{purpose}{what's the purpose of the package being added?}
}
\value{
Nothing. Attempts to modify the packages.R file and prints a
summary message if the modification is successful.
}
\description{
Add a package to packages.R
}
\details{
A purpose isn't necessary, but it can sometimes be helpful for you
or for someone who reads your code. For example, \code{library(here)} is
informative enough - it tells you that the project requires the \code{here}
package. However, if there is a comment explaining that the project
uses the \code{here} package to manage targets in R-markdown documents,
that's a lot more informative.
}
