% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_version.R
\name{assert_valid_version}
\alias{assert_valid_version}
\title{Assert validity of the current project version}
\usage{
assert_valid_version(major, minor, path = "version.rds")
}
\arguments{
\item{major}{(numeric) the current major version of your project}

\item{minor}{(numeric) the current minor version of your project}

\item{path}{a string indicating where \code{version.rds} is stored. This
should \emph{always} be stored in the project's main directory.}
}
\value{
a string representing the current version
}
\description{
This function helps prevent you from accidentally over-writing documents
that you wanted to remain un-modified when you run \code{tar_make()}. For
example, let's say you sent version 1.1 of the results to your colleague
and then you forgot to change version_minor_current to 2. If you make
changes to your code and re-run the pipeline, you'll end up over-writing
version 1.1 of the results. This is annoying and can sometimes create
issues with reproducibility, so \code{assert_valid_version()} is included
at the top of the targets pipeline to make sure this mistake is not
allowed.
}
