% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_helpers.R
\name{use_helpers_datatable}
\alias{use_helpers_datatable}
\title{Add data.table helper functions to R/}
\usage{
use_helpers_datatable()
}
\value{
Nothing. Modifies your workspace.
}
\description{
Copies \code{shift.R} into the \verb{R/} directory of the current project. Contains
\code{shift_forward()} and \code{shift_backward()}, wrappers around
\code{data.table::shift()} that make the shift direction explicit.
}
