% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pericircumference.R
\name{use_pericircumference}
\alias{use_pericircumference}
\title{Initialize pericircumference workflow}
\usage{
use_pericircumference(results_format = "word")
}
\arguments{
\item{results_format}{a character value indicating the format of the
results document. Valid options are
\itemize{
\item "word": use \code{officedown::rdocx_document} results document
}}
}
\value{
Nothing. Modifies your workspace.
}
\description{
Creates files and directories according to the Perisphere template.
The setup is designed to generate tables and figures in the targets
pipeline, then load those objects into a results document. The results
document itself goes through iterations designated by versions. This may
seem unnecessary when using GitHub to handle version control, but it is
often helpful to keep versions of old results documents in a space that
is easy to access.
}
