% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pericircumference_wrappers.R
\name{use_pericircumference_report_ms}
\alias{use_pericircumference_report_ms}
\title{Initialize a Microsoft Word report project}
\usage{
use_pericircumference_report_ms(...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[=use_pericircumference]{use_pericircumference()}},
e.g. \code{report_name = "results"}.}
}
\value{
Nothing. Modifies your workspace.
}
\description{
A wrapper around \code{\link[=use_pericircumference]{use_pericircumference()}} for experienced users who want
to quickly set up a project with a Word report and no slides or tutorials.
Equivalent to calling
\code{use_pericircumference(doc_format = "office", include_tutorials = FALSE, include_report = TRUE, include_slides = FALSE, ...)}.
}
