% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pericircumference_wrappers.R
\name{use_pericircumference_slides_ms}
\alias{use_pericircumference_slides_ms}
\title{Initialize a Microsoft PowerPoint slides project}
\usage{
use_pericircumference_slides_ms()
}
\value{
Nothing. Modifies your workspace.
}
\description{
A wrapper around \code{\link[=use_pericircumference]{use_pericircumference()}} for experienced users who want
to quickly set up a project with PowerPoint slides and no report or
tutorials. Equivalent to calling
\code{use_pericircumference(doc_format = "office", include_tutorials = FALSE, include_report = FALSE, include_slides = TRUE)}.
}
