% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pericircumference_wrappers.R
\name{use_pericircumference_slides_qt}
\alias{use_pericircumference_slides_qt}
\title{Initialize a Quarto HTML slides project}
\usage{
use_pericircumference_slides_qt()
}
\value{
Nothing. Modifies your workspace.
}
\description{
A wrapper around \code{\link[=use_pericircumference]{use_pericircumference()}} for experienced users who want
to quickly set up a project with Quarto slides and no report or tutorials.
Equivalent to calling
\code{use_pericircumference(doc_format = "quarto", include_tutorials = FALSE, include_report = FALSE, include_slides = TRUE)}.
}
